+++ at_secondary/at_secondary_server/CHANGELOG.md
@@ -2,6 +2,7 @@
 - fix: invalidate commit log cache on key deletion
 - feat: remove malformed keys on server startup
 - fix: inbound connection pool test flakiness
+- feat: encode the new line characters in the public key data
 ## 3.0.20
 - fix: Bypass cache rename fix
 - feat: Set isEncrypted to true when notify text message is encrypted.